Removed legacy code (ListViews) for rendering articles.
The reason for moving away for ListViews is that they are too
simple for our usage. Defining an API which listens for both JSON
and HTML gets too messy using the built-in convenience functions,
so we save code writing our own views. This is also more explicit,
and easier to understand for newcomers to the project.