Made the forum link in the navigation link to the facebook group.
Being realistic, we realize we're not going to implement our own forum,
and the treshold for posting in a custom forum here is much higher than
in the facebook group. The Facebook group will now be a dedicated spot
for discussion between our members, while the Facebook page will be our
news channel on Facebook, while the website is our primary resource.