Summer cleanup/restructuring.
Moved all the python source code into it's own package, slingsby. Simplified
the __init__ method considerably, and made sure the WSGI application is defined
here, so that it's easy to import to main.py. This makes it more portable, in case
we would want to experiment with other hosting solutions.

Also threw all the dependencies out of the repo, and specified them in a requirements.txt
file. To set up everything you need for development now, you create a virtualenv called
virtualenv_slingsby, activate it, and do a pip install -r requirements.txt. So much cleaner.

Pulled the gruntfile and package.json out from the build folder and added them to the root
folder, where they should be. Which means you no longer have to be in the build folder to
execute grunt tasks. Puh.

All import statements have been made relative, to enhance portability of the code, I hope
there's no more bugs left because of this, but hard to tell without any automated test suite...

Also made a new grunt task for deploying the static files, and pulled that out of the build task.