Add a general ActionView that can be used for models which has 'actions'.
An action is considered like recieving a vote, being approved, etc. This
view enables such actions to be invoked by URLs such as <id>/vote and
<id>/approve, from the same view.

Accomplished this by overriding the view dispatch-method to look for
a instance member 'actions' which defines legal actions for the model.
In the URLConf you can then link actions to URLs.