Add salt and pillar setups.
SaltStack is a configuration management tool, like Chef and Puppet, only
it's written in Python, and uses YAML + Jinja in config files.

We run a masterless setup (since we're poor), and configure the server
by scp-ing the salt and pillar modules over to the server, and running
salt-call --local on the server. Grunt task to provision and deploy code
is coming soon.

Salt will currently provision the server with postgresql, memcached,
uwsgi, nginx, pip and virtualenv.

Salt will also ensure that the users specified in pillar/users/init.sls
exist on the server and can log in using their SSH key.

Stuff that should be considered for installation later on:
- fail2ban
- iptables
- sendmail