Make Travis deploy static files to fileserver.
Introduces a pillar variable `slingsby_version` that's used for
caching busting. Static files placed in subdirectories on the fileserver,
in directories named after the version. This version is then also
used as part of STATIC_URL, to make sure no cached content is delivered
to a new version.

Note that Travis does not do any housecleaning at the fileserver yet,
so it will fill up. (currently a complete set of static files is ~5MB).

Closes #18.