Use grunt-usemin and grunt-contrib-uglify to minify js.
This also includes some fixes for other build pipeline and grunt related
stuff I discovered while doing this:

- Swap matchdep for load-grunt-tasks.

- Increase readability and diff-friendliness of grunt task aliases. Add
useminPrepare and uglify tasks to build pipeline.

- Bower: Allow minor version upgrade of 1.10 to jquery.

- Move the target of the handlebars grunt task to slingsby/static/.tmp/.
This prevents it from being deployed twice (both minified and normal mode).
Also narrow down the handlebars search path to speed it up a bit.

- Add a argument --port to the grunt server task, so that the portnumber the
devserver starts on can be overridden.

- Ignore the bower libs from being cleaned on every build.

- Fix static file handling (set STATIC_URL='/static/') in the devserver, so
that it finds admin media. Closes #23.

- Add localhost and ntnuita.no to ALLOWED_HOSTS in dev_settings.py so that
slingsby can be tested locally with DEBUG=False.

- Move javascript out of article_list.html and into articles.js. This allows
it to be linted and uglified. Had to add a URL to the js slingsby.urls object
to get rid of a static url tag. This module no longer has to export
improveTimestamp, since the same script is now used on both the article list
and the article details pages, so that was removed.

- Changed collectstatic to happen in two steps, since staticfiles cannot
serve files out of it's target for some reason. So on build, we run
collectstatic to gather all static files in build/.tmp, and then we copy them
over to slingsby/static. Not the most effective solution, but sadly the most
elegant I've found so far.

Eventual people that might happen to read git logs: Sorry for the massive
commit. I know I should have split this up into at least 3-4 commits. But
I'm excercising my right to not give a fuck since we have no stakeholders
that care about the git log, except me.