Better secret handling
Store secrets encrypted in salt/secure/init.sls, and provide a tool to
encrypt/decrypt new secrets. Key is stored in Kontoer.kdbx in the webkom
dropbox.

This makes sure that the only thing needed for a new dev to be able to do
anything, is to know the password to the keepass db in the dropbox, then
you can run devserver locally with all secrets needed, like facebook api
tokens and ssh keys to the fileserver (will be needed soon).