Don't reload on exceptions
This is what's been causing stuff to crash on heavy signup load. The idea
is that if the application crashes, restart the interpreter so that if the
crash was caused by stale variables, memory errors or anything temporary,
a restart would clear it out. During load testing this didn't cause any
problems, as the load tester always performed complete HTTP requests.
However, during signups, people go mad on the F5 key and re-try queries,
causing the current request to be aborted, and triggering a write error
for uwsgi trying to write to a closed socket, which triggered a uwsgi
restart. When most of the requests are aborted (because when uwsgi first
has restarted, it'll be slow for the next user, causing him to abort the
request and fire off a new one), this causes uwsgi to enter a continious
restart loop, and causing the entire site to be mostly unavailable during
signups.